ftp condicion de si existe archivo

+++ laravel/app/controllers/RutaDetalleController.php
@@ -835,13 +835,14 @@ writen by Jhoubert @ Veflat.com
 
     function getRefreshfiles(){
         $files = $this->prepareFiles();
-        FtpFiles::truncate();
-        foreach ($files as $ind => $tFile) {
-            $files = new FtpFiles;
-            $files['link']=$tFile;
-            $files->save();
+        if(is_array($files) && count($files)>0){
+            FtpFiles::truncate();
+            foreach ($files as $ind => $tFile) {
+                $files = new FtpFiles;
+                $files['link']=$tFile;
+                $files->save();
+            }
         }
-
         echo "OK"; 
     }
